refactor(recipe): split rules_campaign.py into 5 focused submodules
Split the 1076-line rules_campaign.py into 5 thematically-cohesive
submodules following the "one file per semantic-rule domain" convention:

- rules_campaign_dispatch.py: campaign structure, naming, recipe refs (9 rules)
- rules_campaign_ingredients.py: ingredient validation (4 rules)
- rules_campaign_deps.py: dependency graph validation (3 rules)
- rules_campaign_capture.py: capture/sentinel validation (4 rules)
- rules_campaign_flow.py: gates, paths, version, skip-when (9 rules)

Extracted _extract_sentinel_fields and _load_dispatch_target to
_rule_helpers.py as shared helpers consumed by multiple submodules.

Updated:
- recipe/__init__.py: replaced single rules_campaign import with 5 new imports
- tests/_test_filter.py: replaced rules_campaign cascade entry with 5 new entries
- tests/arch/test_subpackage_isolation.py: removed line-limit exemption,
  raised rules/ file count ceiling from 33 to 37
- tests/recipe/test_rules_campaign.py, test_bundled_recipes_dispatch_ready.py,
  test_campaign_loader.py: updated imports to new locations
- docs and CLAUDE.md files: updated file counts and cross-references
