+++ tests/_test_filter.py
@@ -357,7 +357,11 @@ MODULE_CASCADE_RECIPE: dict[str, frozenset[str]] = {
     "rules_actions": frozenset({"recipe"}),
     "rules_blocks": frozenset({"recipe"}),
     "rules_bypass": frozenset({"recipe"}),
-    "rules_campaign": frozenset({"recipe"}),
+    "rules_campaign_capture": frozenset({"recipe"}),
+    "rules_campaign_deps": frozenset({"recipe"}),
+    "rules_campaign_dispatch": frozenset({"recipe"}),
+    "rules_campaign_flow": frozenset({"recipe"}),
+    "rules_campaign_ingredients": frozenset({"recipe"}),
     "rules_ci": frozenset({"recipe"}),
     "rules_ci_conflict": frozenset({"recipe"}),
     "rules_ci_guards": frozenset({"recipe"}),
